articles/01-text-embeddings/requirements.txt
mteb
sentence-transformers
PyYAML
